fix: recalculate client aggregates when override flags change
setServiceOverride, setServiceOverdraftResolved, and moveToNextStage
now call calcClientAggregates inside their transactions, so isBlocked
and isCritical are immediately correct after toggling overrides.

Extract calcClientAggregates to shared/aggregates.js — single source
of truth for both clients/ and services/ modules. Override-aware:
respects overrideActive and overdraftResolved.isResolved.
